dairy-demo-bot:server/requirements.txt
asn1crypto==0.24.0
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
future==0.17.1
google-api-core==1.11.1
google-auth==1.6.3
google-cloud-speech==1.0.0
googleapis-common-protos==1.6.0
grpcio==1.21.1
idna==2.8
nltk==3.4.1
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
python-telegram-bot==11.1.0
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
SpeechRecognition==3.8.1
textblob==0.15.3
urllib3==1.25.3
python-telegram-bot==11.1.0
psycopg2==2.8.3
rake-nltk==1.0.4
tensorflow==1.14.0
keras==2.2.4
